% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_DESeq.R
\name{create_DESeq}
\alias{create_DESeq}
\title{Create DESeq object or list of DESeq objects}
\usage{
create_DESeq(
  pseudobulk_counts,
  pseudobulk_metadata,
  counts_filter,
  Design,
  split = FALSE
)
}
\arguments{
\item{pseudobulk_counts}{Data frame of gene by sample counts}

\item{pseudobulk_metadata}{Data frame of sample metadata}

\item{counts_filter}{Int minimum count to retain gene}

\item{Design}{Formula for model to fit}

\item{split}{Lgl is pseudobulk data split by celltype}
}
\value{
DESeq object, or list of DESeq objects if \code{split == TRUE}
}
\description{
Create DESeq object or list of DESeq objects
}
\examples{
# create_DESeq(
#     pseudobulk_counts = counts.list,
#   pseudobulk_metadata = metadata.list,
#   counts_filter = 5,
#   Design = ~Treatment,
#   split = TRUE
# )

}
