% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_metadata.R
\name{create_metadata}
\alias{create_metadata}
\title{Create Pseudo-bulked metadata table}
\usage{
create_metadata(
  seurat_obj,
  sample_id_vars,
  metadata_vars,
  cell_type_var,
  split = FALSE
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{sample_id_vars}{Chr vector of metadata fields to construct sample ID from}

\item{metadata_vars}{Chr vector of metadata fields to include in metadata output}

\item{cell_type_var}{Chr name of metadata field including cell type annotation}

\item{split}{Lgl weather or not metadata table should be split by cell type}
}
\value{
Data frame, list of data frames if \code{split == TRUE}
}
\description{
Create Pseudo-bulked metadata table
}
\examples{
create_metadata(
    seurat_obj = Neurons,
    sample_id_vars = c("Sex", "GEM"),
    metadata_vars = c("Sex", "GEM", "Treatment"),
    cell_type_var = "CellType",
    split = TRUE
)

}
