% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.create.safe.R
\name{create_package0}
\alias{create_package0}
\title{Old Devtools create_package}
\usage{
create_package0(path, description = getOption("devtools.desc"),
  check = FALSE, rstudio = TRUE, quiet = FALSE)
}
\description{
Old Devtools create_package
}
