% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.create.safe.R
\name{dir.create.safe}
\alias{dir.create.safe}
\title{Create A New Directory}
\usage{
dir.create.safe(DIR)
}
\arguments{
\item{DIR}{a directory}
}
\value{
nothing
}
\description{
Create A New Directory
}
\examples{
dir.create.safe( "/tmp/Trial" )
}
