% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_above.R
\name{fill_above}
\alias{fill_above}
\title{Trim Whitespace from a String}
\usage{
fill_above(x, blank = is.na)
}
\arguments{
\item{x}{a vector}

\item{blank}{the blank value to be replaced by the row above}
}
\value{
a vector
}
\description{
Trim Whitespace from a String
}
\examples{
fill_above( c("A", "C", "", "B") , blank="")
}
