% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_na.R
\name{list_nas}
\alias{list_nas}
\title{Which list elements are NA}
\usage{
list_nas(LIST)
}
\arguments{
\item{LIST}{list}
}
\value{
vector indicating which elements are not na
}
\description{
Which list elements are NA
}
