% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgtemplate-package.R
\docType{package}
\name{miscutils-package}
\alias{miscutils}
\alias{miscutils-package}
\title{miscutils: Miscellaneous utility functions}
\description{
Miscellaneous utility functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/audrey-b/miscutils}
  \item Report bugs at \url{https://github.com/audrey-b/miscutils/issues}
}

}
\author{
\strong{Maintainer}: Audrey Beliveau \email{audrey.beliveau@uwaterloo.ca}

}
\keyword{internal}
