% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.create.safe.R
\name{package.setup.safe}
\alias{package.setup.safe}
\title{Create A New R Package}
\usage{
package.setup.safe(DIR, my_description = getOption("devtools.desc"),
  rstudio = FALSE, check = FALSE, rm = FALSE)
}
\arguments{
\item{DIR}{a directory}
}
\value{
nothing
}
\description{
Create A New R Package
}
\examples{
dir.create.safe( "/tmp/Trial" )
}
