% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dim_gene.R
\name{plot_dim_geme}
\alias{plot_dim_geme}
\title{Plot by cell gene expression}
\usage{
plot_dim_geme(SCE, Dim, Gene, Exp_Order = FALSE)
}
\arguments{
\item{SCE}{Single Cell Experiment object}

\item{Dim}{String. Name of dimensionality reduciton to plot}

\item{Gene}{String. Name of gene to plot}

\item{Exp_Order}{Boolean. If TRUE, cells will be plotted in order of count values}
}
\value{
A ggplot object
}
\description{
Plot gene expression (in log counts) for a particular gene within cells in
a reduced dimensionality plot. This function is borrowed in large part from
the iSEE reduced dimensionality plot panel.
}
\examples{
# marker_plot(SCE = se,
#             Dim = "PC12UMAP",
#             Gene = "Th",
#             Exp_Order = TRUE
# )
}
