% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_integration_bar.R
\name{plot_integration_bar}
\alias{plot_integration_bar}
\title{Plot cluster integration as stacked bar plot}
\usage{
plot_integration_bar(seurat_obj, cluster_col, treatment_col, gem_col, sex_col)
}
\arguments{
\item{sex_col}{}
}
\description{
Plot cluster integration as stacked bar plot
}
