% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render2.R
\name{render2}
\alias{render2}
\title{A function to render and preview markdown documents.}
\usage{
render2(input = rstudioapi::getSourceEditorContext()$path,
  replace = getOption("path.replace.list"), preview = TRUE, ...)
}
\arguments{
\item{input}{The path of the input file to be rendered. Generally an R Markdown document (.Rmd). Defaults to the active file in the RStudio editor.}

\item{replace}{A list of two strings. The first string will be replaced by the second one in the path. This can be used to solve issues when working from a mapped disk.}

\item{preview}{Whether to preview the rendered documents.}

\item{...}{Arguments to pass to \code{\link[rmarkdown]{render}}. For example, output_format="all" will render all formats defined within the file.}
}
\description{
This function provides an alternative to the Knit button in RStudio, which tends to fail when working from a mapped disk. 
It is based on the \code{\link[rmarkdown]{render}} function but also provides a preview of the rendered document.
}
