% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_DESeq.R
\name{results_DESeq}
\alias{results_DESeq}
\title{Get DESeq Results}
\usage{
results_DESeq(DESeq_obj, seurat_obj, Contrast)
}
\arguments{
\item{DESeq_obj}{DESeq object}

\item{seurat_obj}{Seurat object}

\item{Contrast}{Chr vector of contrast in the format c("variable", "level1", "level2")}
}
\value{
Data frame of DESeq results
}
\description{
Get DESeq Results
}
\examples{
# results_DESeq(
#   seurat_obj = Neurons,
#   DESeq_obj = DESeq_obj,
#   Contrast = c("Treatment", "Mor", "Sal")
# )

}
