% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmall.R
\name{rmall}
\alias{rmall}
\title{remove all objects}
\usage{
rmall(ENVIR = globalenv(), LIST = ls(envir = ENVIR), EXCEPT = NULL)
}
\arguments{
\item{ENVIR}{environmnet to remove from}

\item{LIST}{strings to remove}

\item{EXCEPT}{vector of strings for any exceptions to keep}
}
\value{
nothing
}
\description{
remove all objects
}
