% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range0.R
\name{round00}
\alias{round00}
\alias{ceiling00}
\alias{floor00}
\title{Returns Floor and Ceilings}
\usage{
round00(x, e = 0.25, ...)

ceiling00(x, e = 0.25)

floor00(x, e = 0.25)
}
\arguments{
\item{x}{a numeric vector}

\item{e}{decimal rounding (.25) means nearest quarters}
}
\description{
Returns Floor and Ceilings
}
\seealso{
round00 ceiling0

round00 floor0
}
