% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.create.safe.R
\name{setup0}
\alias{setup0}
\title{Old Devtools setup}
\usage{
setup0(path = ".", description = getOption("devtools.desc"),
  check = FALSE, rstudio = TRUE, quiet = FALSE)
}
\description{
Old Devtools setup
}
