% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trim.R
\name{trim.all}
\alias{trim.all}
\title{returns string without leading or trailing whitespace}
\usage{
trim.all(x, s = "")
}
\arguments{
\item{x}{A character string}
}
\description{
returns string without leading or trailing whitespace
}
