% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trim.R
\name{trim.trailing}
\alias{trim.trailing}
\title{returns string without trailing whitespace}
\usage{
trim.trailing(x, r = "\\\\s", s = "")
}
\arguments{
\item{x}{A character string}

\item{s, r}{alternative replacement character}
}
\description{
returns string without trailing whitespace
}
