% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wideScreen.R
\name{wideScreen}
\alias{wideScreen}
\title{Format Terminal Viewing Screen}
\usage{
wideScreen(howWide = Sys.getenv("COLUMNS"), howTall = Sys.getenv("LINES"),
  tt = FALSE)
}
\arguments{
\item{howWide, howTall}{number of terminal columns and rows}
}
\description{
Format Terminal Viewing Screen
}
