% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/within_celltype_DEG.R
\name{within_celltype_DEG}
\alias{within_celltype_DEG}
\title{Conduct Pseudo-bulked Differential Expression Within Cell Types}
\usage{
within_celltype_DEG(
  seurat_obj,
  cell_type_var,
  grouping_vars,
  counts_filter = 5,
  Design,
  Contrast,
  ...
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{cell_type_var}{Chr name of metadata variable containing cell type annotation}

\item{grouping_vars}{Chr vector of metadata variables that define cell grouping strategy}

\item{counts_filter}{Int minimum count to retain gene}

\item{Design}{Formula for model to fit}

\item{Contrast}{Chr vector of contrast in the format c("variable", "level1", "level2")}

\item{...}{Additional arguments to pass to DESeq2}
}
\value{
Data frame of DESeq results
}
\description{
This function conducts our standard workflow for pseudo-bulked differential
gene expression testing within cell types using DESeq2. For a more modular
and interactive workflow, see the individual \code{create_pseudobulk()},
\code{create_metadata()}, \code{create_DESeq()}, \code{DESeq()}, and \code{results_DESeq()} functions.
}
\examples{
within_celltype_DEG(
    seurat_obj = Neurons,
    cell_type_var = "CellType",
    grouping_vars = c("Sex", "GEM"),
    Design = ~Treatment,
    Contrast = c("Treatment", "Mor", "Sal")
)
}
