% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_integration_lisi.R
\name{plot_integration_lisi}
\alias{plot_integration_lisi}
\title{Plot cluster integration by LISI as ridge plot}
\usage{
plot_integration_lisi(
  seurat_obj,
  cluster_col,
  treatment_col,
  gem_col,
  sex_col,
  reduction = "umap"
)
}
\arguments{
\item{reduction}{}
}
\description{
Plot cluster integration by LISI as ridge plot
}
