% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range0.R
\name{range0}
\alias{range0}
\alias{floor0}
\alias{ceiling0}
\alias{seq0}
\title{Returns Floor and Ceilings}
\usage{
range0(x, e = 1, keepna = T)

floor0(x, e = 1)

ceiling0(x, e = 1)

seq0(x, e = 1, ...)
}
\arguments{
\item{x}{a numeric vector}

\item{e}{decimal place to round to}

\item{keepna}{logical keep NA values}
}
\description{
Returns Floor and Ceilings
}
\seealso{
range0 ceiling00

range0 ceiling00
}
