% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render2.R
\name{render2}
\alias{render2}
\title{A function to render and preview markdown documents.}
\usage{
render2(filename, dir = getwd(), output_format = "pdf_document", ...)
}
\arguments{
\item{filename}{The name of the input file to be rendered. Generally an R Markdown document (.Rmd).}

\item{dir}{The directory where the input file is located.}

\item{output_format}{The R Markdown output format to convert to. Currently only "pdf_document" and "html_document" are implemented.}

\item{...}{Arguments to pass to rmarkdown::render.}
}
\description{
This function provides a solution to the knit button in RStudio, which fails when working from a mounted disk. It has the
advantage over the rmarkdown::render function to provide a preview of the rendered document.
}
