% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trim.R
\name{trim.leading}
\alias{trim.leading}
\title{returns string without leading whitespace}
\usage{
trim.leading(x, r = "\\\\s", s = "")
}
\arguments{
\item{x}{A character string}

\item{s, r}{alternative replacement character}
}
\description{
returns string without leading whitespace
}
