% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trim.R
\name{trim}
\alias{trim}
\title{Trim Whitespace from a String}
\usage{
trim(x, s = "", strip = "all")
}
\arguments{
\item{x}{A character string}

\item{strip}{the whitespace to be removed: 'trailing', 'leading', or 'all'}
}
\value{
A character string
}
\description{
Trim Whitespace from a String
}
\examples{
trim( " AA   ")
}
